+++ src/Aws/ElasticTranscoder/Resources/elastictranscoder-2012-09-25.php
@@ -1107,10 +1107,22 @@ return array (
                         'Channels' => array(
                             'type' => 'string',
                         ),
+                        'AudioPackingMode' => array(
+                            'type' => 'string',
+                        ),
                         'CodecOptions' => array(
                             'type' => 'object',
                             'properties' => array(
                                 'Profile' => array(
+                                    'type' => 'string',
+                                ),
+                                'BitDepth' => array(
+                                    'type' => 'string',
+                                ),
+                                'BitOrder' => array(
+                                    'type' => 'string',
+                                ),
+                                'Signed' => array(
                                     'type' => 'string',
                                 ),
                             ),
@@ -2708,10 +2720,22 @@ return array (
                                 'Channels' => array(
                                     'type' => 'string',
                                 ),
+                                'AudioPackingMode' => array(
+                                    'type' => 'string',
+                                ),
                                 'CodecOptions' => array(
                                     'type' => 'object',
                                     'properties' => array(
                                         'Profile' => array(
+                                            'type' => 'string',
+                                        ),
+                                        'BitDepth' => array(
+                                            'type' => 'string',
+                                        ),
+                                        'BitOrder' => array(
+                                            'type' => 'string',
+                                        ),
+                                        'Signed' => array(
                                             'type' => 'string',
                                         ),
                                     ),
@@ -4465,10 +4489,22 @@ return array (
                                     'Channels' => array(
                                         'type' => 'string',
                                     ),
+                                    'AudioPackingMode' => array(
+                                        'type' => 'string',
+                                    ),
                                     'CodecOptions' => array(
                                         'type' => 'object',
                                         'properties' => array(
                                             'Profile' => array(
+                                                'type' => 'string',
+                                            ),
+                                            'BitDepth' => array(
+                                                'type' => 'string',
+                                            ),
+                                            'BitOrder' => array(
+                                                'type' => 'string',
+                                            ),
+                                            'Signed' => array(
                                                 'type' => 'string',
                                             ),
                                         ),
@@ -5495,10 +5531,22 @@ return array (
                                 'Channels' => array(
                                     'type' => 'string',
                                 ),
+                                'AudioPackingMode' => array(
+                                    'type' => 'string',
+                                ),
                                 'CodecOptions' => array(
                                     'type' => 'object',
                                     'properties' => array(
                                         'Profile' => array(
+                                            'type' => 'string',
+                                        ),
+                                        'BitDepth' => array(
+                                            'type' => 'string',
+                                        ),
+                                        'BitOrder' => array(
+                                            'type' => 'string',
+                                        ),
+                                        'Signed' => array(
                                             'type' => 'string',
                                         ),
                                     ),