+++ src/Aws/Kinesis/Resources/kinesis-2013-12-02.php
@@ -893,6 +893,10 @@ return array (
                     'type' => 'string',
                     'location' => 'json',
                 ),
+                'MillisBehindLatest' => array(
+                    'type' => 'numeric',
+                    'location' => 'json',
+                ),
             ),
         ),
         'GetShardIteratorOutput' => array(